permission permission check done

+++ app/Http/Controllers/Backend/Admin/PermissionController.php
@@ -20,6 +20,7 @@ class PermissionController extends Controller
      */
     public function index(Request $request)
     {
+        // if (auth()->user()->can('permission_read')) {
         $name = null;
         $permissions = AdminPermission::latest();
         if ($request->has('name') && !empty($request->input('name'))) {
@@ -29,6 +30,10 @@ class PermissionController extends Controller
         $permissions = $permissions->paginate(15);
 
         return view('backend.admin.permission.index', compact('permissions', 'name'));
+        // } else {
+        //     $link = "admin.dashboard";
+        //     return view('error.403', compact('link'));
+        // }
     }
 
     /**
@@ -37,9 +42,13 @@ class PermissionController extends Controller
     public function create(Request $request)
     {
         if ($request->ajax()) {
-            $modules = Module::whereStatus(1)->whereIsLabel(0)->get();
-            $view = View::make('backend.admin.permission.create', compact('modules'))->render();
-            return response()->json(['html' => $view]);
+            if (auth()->user()->can('permission_create')) {
+                $modules = Module::whereStatus(1)->whereIsLabel(0)->get();
+                $view = View::make('backend.admin.permission.create', compact('modules'))->render();
+                return response()->json(['html' => $view]);
+            } else {
+                abort(403, 'Sorry, you are not authorized to access this page');
+            }
         } else {
             return response()->json(['status' => 'false', 'message' => "Access only ajax request"]);
         }
@@ -51,45 +60,48 @@ class PermissionController extends Controller
     public function store(Request $request)
     {
         if ($request->ajax()) {
+            if (auth()->user()->can('permission_create')) {
+                $rules = [
+                    'module_id' => 'required',
+                    'name' => 'required',
+                    'label' => 'required',
+                ];
 
-            $rules = [
-                'module_id' => 'required',
-                'name' => 'required',
-                'label' => 'required',
-            ];
-
-            $validator = Validator::make($request->all(), $rules);
-            if ($validator->fails()) {
-                return response()->json([
-                    'type' => 'error',
-                    'errors' => $validator->getMessageBag()->toArray()
-                ]);
-            } else {
-                $modules = $request->input('module_id');
+                $validator = Validator::make($request->all(), $rules);
+                if ($validator->fails()) {
+                    return response()->json([
+                        'type' => 'error',
+                        'errors' => $validator->getMessageBag()->toArray()
+                    ]);
+                } else {
+                    $modules = $request->input('module_id');
 
 
-                DB::beginTransaction();
-                try {
-                    foreach ($modules as $module) {
+                    DB::beginTransaction();
+                    try {
+                        foreach ($modules as $module) {
 
-                        $arr = explode(', ', $module);
+                            $arr = explode(', ', $module);
 
-                        $data = [
-                            'name' => $arr[1] . '_' . $request->input('name'),
-                            'label' => $request->input('label'),
-                            'module_id' => $arr[0],
-                            'is_visibile_to_role' => $request->has('is_visibile_to_role') ? 1 : 0
-                        ];
-                        Permission::Create($data);
+                            $data = [
+                                'name' => $arr[1] . '_' . $request->input('name'),
+                                'label' => $request->input('label'),
+                                'module_id' => $arr[0],
+                                'is_visibile_to_role' => $request->has('is_visibile_to_role') ? 1 : 0
+                            ];
+                            Permission::Create($data);
+                        }
+
+                        DB::commit();
+
+                        return response()->json(['type' => 'success', 'message' => "Successfully Created"]);
+                    } catch (Exception $e) {
+                        DB::rollback();
+                        return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
                     }
-
-                    DB::commit();
-
-                    return response()->json(['type' => 'success', 'message' => "Successfully Created"]);
-                } catch (Exception $e) {
-                    DB::rollback();
-                    return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
                 }
+            } else {
+                abort(403, 'Sorry, you are not authorized to access this page');
             }
         } else {
             return response()->json(['status' => 'false', 'message' => "Access only ajax request"]);
@@ -102,8 +114,12 @@ class PermissionController extends Controller
     public function show(Request $request, AdminPermission $permission)
     {
         if ($request->ajax()) {
-            $view = View::make('backend.admin.permission.view', compact('permission'))->render();
-            return response()->json(['html' => $view]);
+            if (auth()->user()->can('permission_read')) {
+                $view = View::make('backend.admin.permission.view', compact('permission'))->render();
+                return response()->json(['html' => $view]);
+            } else {
+                abort(403, 'Sorry, you are not authorized to access this page');
+            }
         } else {
             return response()->json(['status' => 'false', 'message' => "Access only ajax request"]);
         }
@@ -115,9 +131,13 @@ class PermissionController extends Controller
     public function edit(Request $request, Permission $permission)
     {
         if ($request->ajax()) {
-            $modules = Module::whereIsLabel(0)->get();
-            $view = View::make('backend.admin.permission.edit', compact('permission', 'modules'))->render();
-            return response()->json(['html' => $view]);
+            if (auth()->user()->can('permission_update')) {
+                $modules = Module::whereIsLabel(0)->get();
+                $view = View::make('backend.admin.permission.edit', compact('permission', 'modules'))->render();
+                return response()->json(['html' => $view]);
+            } else {
+                abort(403, 'Sorry, you are not authorized to access this page');
+            }
         } else {
             return response()->json(['status' => 'false', 'message' => "Access only ajax request"]);
         }
@@ -129,38 +149,41 @@ class PermissionController extends Controller
     public function update(Request $request, Permission $permission)
     {
         if ($request->ajax()) {
+            if (auth()->user()->can('permission_update')) {
+                $rules = [
+                    'module_id' => 'required',
+                    'name' => 'required|unique:permissions,name,' . $permission->id,
+                    'label' => 'required',
+                ];
 
-            $rules = [
-                'module_id' => 'required',
-                'name' => 'required|unique:permissions,name,' . $permission->id,
-                'label' => 'required',
-            ];
+                $validator = Validator::make($request->all(), $rules);
+                if ($validator->fails()) {
+                    return response()->json([
+                        'type' => 'error',
+                        'errors' => $validator->getMessageBag()->toArray()
+                    ]);
+                } else {
 
-            $validator = Validator::make($request->all(), $rules);
-            if ($validator->fails()) {
-                return response()->json([
-                    'type' => 'error',
-                    'errors' => $validator->getMessageBag()->toArray()
-                ]);
-            } else {
+                    DB::beginTransaction();
+                    try {
 
-                DB::beginTransaction();
-                try {
+                        $permission->module_id = $request->input('module_id');
+                        $permission->name = $request->input('name');
+                        $permission->label = $request->input('label');
+                        $permission->is_visibile_to_role = $request->has('is_visibile_to_role') ? 1 : 0;
+                        $permission->status = $request->input('status');
+                        $permission->save();
 
-                    $permission->module_id = $request->input('module_id');
-                    $permission->name = $request->input('name');
-                    $permission->label = $request->input('label');
-                    $permission->is_visibile_to_role = $request->has('is_visibile_to_role') ? 1 : 0;
-                    $permission->status = $request->input('status');
-                    $permission->save();
+                        DB::commit();
 
-                    DB::commit();
-
-                    return response()->json(['type' => 'success', 'message' => "Successfully Updated"]);
-                } catch (Exception $e) {
-                    DB::rollback();
-                    return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+                        return response()->json(['type' => 'success', 'message' => "Successfully Updated"]);
+                    } catch (Exception $e) {
+                        DB::rollback();
+                        return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+                    }
                 }
+            } else {
+                abort(403, 'Sorry, you are not authorized to access this page');
             }
         } else {
             return response()->json(['status' => 'false', 'message' => "Access only ajax request"]);
@@ -173,15 +196,18 @@ class PermissionController extends Controller
     public function destroy(Request $request, AdminPermission $permission)
     {
         if ($request->ajax()) {
-
-            DB::beginTransaction();
-            try {
-                $permission->delete();
-                DB::commit();
-                return response()->json(['type' => 'success', 'message' => 'Successfully Deleted']);
-            } catch (Exception $e) {
-                DB::rollback();
-                return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+            if (auth()->user()->can('permission_delete')) {
+                DB::beginTransaction();
+                try {
+                    $permission->delete();
+                    DB::commit();
+                    return response()->json(['type' => 'success', 'message' => 'Successfully Deleted']);
+                } catch (Exception $e) {
+                    DB::rollback();
+                    return response()->json(['type' => 'error', 'message' => "<div class='alert alert-danger'>" . $e->getMessage() . "</div>"]);
+                }
+            } else {
+                abort(403, 'Sorry, you are not authorized to access this page');
             }
         } else {
             return response()->json(['status' => 'false', 'message' => "Access only ajax request"]);
